Move utility functions from MCXcor module to more appropriate locations
